pyproject.toml
[project]
name = "impresora-termica"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Flet developer", email = "you@example.com" }
]
dependencies = [
  "flet==0.27.1",
  "pillow==10.4.0"
  
]

[tool.flet]
# org name in reverse domain name notation, e.g. "com.mycompany".
# Combined with project.name to build bundle ID for iOS and Android apps
org = "com.mycompany"


# project display name that is used as an app title on Android and iOS home screens,
# shown in window titles and about app dialogs on desktop.
product = "impresora-termica"

# company name to display in about app dialogs
company = "Flet"

# copyright text to display in about app dialogs
copyright = "Copyright (C) 2025 by Flet"

[tool.flet.app]
path = "src"
compile_app = true
compile_packages = true
cleanup_on_compile = true

[tool.uv]
dev-dependencies = [
    "flet[all]==0.27.1",
]

[tool.poetry]
package-mode = true

[tool.poetry.group.dev.dependencies]
flet = {extras = ["all"], version = "0.27.1"}